Move contributors plugin/tasks to appropriate package
