+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.contributors.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.ContributorsPlugin
+implementation-class=org.shipkit.internal.gradle.contributors.ContributorsPlugin
